fix: refactor message to follow app structure and result convention
