+++ app/src/main/java/com/android/bookswap/data/DataMessage.kt
@@ -1,0 +1,9 @@
+package com.android.bookswap.data
+
+data class DataMessage(
+    val id: String = "",
+    val text: String = "",
+    val senderId: String = "",
+    val receiverId: String = "",
+    val timestamp: Long = 0L
+)
