--- app/src/main/java/com/android/bookswap/data/chat/MessageViewModel.kt
@@ -1,9 +1,0 @@
-package com.android.bookswap.data.chat
-
-data class Message(
-    val id: String = "",
-    val text: String = "",
-    val senderId: String = "",
-    val receiverId: String = "",
-    val timestamp: Long = 0L
-)
